+++ src/math/extremesearching.h
@@ -11,7 +11,7 @@ private:
 
   std::fstream debug;
 
-  long QCount, _I_, IMax, IMin, Count, I_end;
+  int QCount, _I_, IMax, IMin, Count, I_end;
   double YMax, YMin;
 
   double CalcNoise(void);
@@ -19,27 +19,27 @@ private:
   int CheckForCentralExtreme();
 
 protected:
-  virtual void OnMaximum(double Value, long Index) = 0;
-  virtual void OnMinimum(double Value, long Index) = 0;
+  virtual void OnMaximum(double Value, int Index) = 0;
+  virtual void OnMinimum(double Value, int Index) = 0;
 
 public:
   class TDataHandler
   {
   public:
-    virtual double GetX(long index) = 0;
-    virtual double GetY(long index) = 0;
-    virtual long Count(void) = 0;
+    virtual double GetX(int index) = 0;
+    virtual double GetY(int index) = 0;
+    virtual int Count(void) = 0;
   };
 
   TDataHandler *Data;
-  long ChainPoints;
+  int ChainPoints;
   double Noise;
   bool LeftBoundary;
   bool RightBoundary;
 
   TExtremeSearcher(
     TDataHandler *Data = nullptr,
-    long ChainPoints   = 3,
+    int ChainPoints   = 3,
     double Noise       = 0,
     bool LeftBoundary  = false,
     bool RightBoundary = false
